#1153 - Move report skill notifications, metric logging and user lookup outside of DB transaction
